automatically backed up by learn

+++ db/migrate/001_create_costumes.rb
@@ -12,7 +12,7 @@ class CreateCostumes < ActiveRecord::Mirgration[4.2]
   def change
     create_table :costumes do |c|
       c.string :name
-      c.float :price
+      c.integer :price
       c.string :size
       c.string :image_url
       c.timestamps
